fixes test suite inside current module structure
