+++ src/main.rs
@@ -1,4 +1,5 @@
-pub mod binary01;
+mod binary01;
+mod binary01_tests;
 
 fn main() {
     println!("Hello, world!");
